test github action script to update-rust-components
